+++ CHANGES.md
@@ -1,3 +1,8 @@
+2011.12.17, Version 1.0.2
+
+* [#1](https://github.com/neocotic/mor.js/issues/1): Add support for single string or object argument to the `encode` and `decode` methods
+* [#3](https://github.com/neocotic/mor.js/issues/3): Expose read-only versions of the internal `chars` and `modes` variables
+
 2011.11.10, Version 1.0.1
 
 * Add support for [Node.js](http://nodejs.org), [CommonJS](http://commonjs.org) and [Ender](http://ender.no.de)